Refactor for shims 320+
